Implement 64 bit audio processing and AVX support
Uses double float processing across the entire pipeline within EAPO for improved precision when overlaying effects.

Also uses AVX256 and AVX512, along with FMA instructions to ensure maximum performance, with AVX512 offering a theoretical 2x performance uplift on say a Zen 5 processor.

+++ IFilter.h
@@ -38,7 +38,7 @@ public:
 	virtual bool getSelectChannels() {return false;}
 	// return value is the channelNames vector, which may contain additional or fewer channel names
 	virtual std::vector<std::wstring> initialize(float sampleRate, unsigned maxFrameCount, std::vector<std::wstring> channelNames) = 0;
-	virtual void process(float** output, float** input, unsigned frameCount) = 0;
+	virtual void process(double** output, double** input, unsigned frameCount) = 0;
 
 protected:
 };
