+++ filters/ConvolutionFilter.h
@@ -30,7 +30,7 @@ public:
 	virtual ~ConvolutionFilter();
 	bool getInPlace() override { return true; }
 	std::vector<std::wstring> initialize(float sampleRate, unsigned maxFrameCount, std::vector<std::wstring> channelNames) override;
-	void process(float** output, float** input, unsigned frameCount) override;
+	void process(double** output, double** input, unsigned frameCount) override;
 
 protected:
 	virtual void initializeFilters(unsigned frameCount);
